refactor: consolidate into `sql-identifier-quoting.ts`
